Code refactor and added missing docs
