+++ .swiftlint.yml
@@ -16,6 +16,7 @@ missing_docs:
     - internal
     - public
     - open
+
 opt_in_rules:
   - attributes
   - balanced_xctest_lifecycle
