--- KMART/Model/GroupType.swift
@@ -1,14 +1,0 @@
-//
-//  GroupType.swift
-//  KMART
-//
-//  Created by Nindi Gill on 27/2/21.
-//
-
-import Foundation
-
-enum GroupType: String {
-    case common = "Common"
-    case mac = "Mac"
-    case mobile = "Mobile"
-}
